requirements.txt
pytest==8.0.0
selene==2.0.0rc7
allure-pytest==2.9.45
python-dotenv==0.20.0